[SOLARIS] Add atomic.h to system-headers.
https://bugzilla.mozilla.org/show_bug.cgi?id=1369061

OpenIndiana used a much messier fix for this, but this one was used by Mozilla and looks a lot cleaner. It shouldn't interfere with any other targets, but if it does, the messy version of this fix basically involves rewriting the Solaris version of atomicops_internals_solaris.h to use GCC compiler intrinsics for atomic operations directly. It absolutely works, but it's gross to look at.

https://github.com/OpenIndiana/oi-userland/blob/3b246b0b385735b092a88f58f9baa9799ee34761/components/web/firefox/patches/01-FF43.0b3_OpenSXCE_x86_x64.patch

Another fix may be possible by preventing config/gcc-hidden.h from being included, or possibly using well-placed GCC pragmas to solve the visibility issues.