Issue #1332 - Test promise-based media playback
https://bugzilla.mozilla.org/show_bug.cgi?id=1244768

I did this patch yesterday because I wanted a quick way of testing whether this promise implementation would work with UXP. Mozilla had changed too much to make this quick and easy, so I instead backported all of the UXP changes to the newer version of the file.

The only real backporting I did was the bare minimum to get it to compile, undoing a few minor DocShell changes. This may very well break something important, but at least we know it can be made to work.