Repair fallout from Bug 1426494
I'm not sure what exactly happened to cause the circular header dependency issues, but I seem to have resolved them anyway. I also wound up needing the devirtualize part of the patch because the way the "Share more code" part is structured depends on it.

As for the rest of the bug, here is how I got this working. Look closely at the original bug and notice that it's moving code around pretty much intact, but the way the patch was applied here removes old code and replaces it with much newer Firefox code that hasn't been adapted to UXP.

https://bugzilla.mozilla.org/show_bug.cgi?id=1217436

Specifically we didn't have this, and all it did was change an nsTArray to an AutoTArray and added no new functionality whatsoever. So I accounted for us not having it. It makes it really obvious what the missing callbacks/prototypes are when you understand what this patch did and how we got half of it inadvertantly when moving code between nsIdentifierMapEntry and nsDocument.

https://bugzilla.mozilla.org/show_bug.cgi?id=1355787

This is where AtomOrString comes from, and it appears we don't technically need this patch for this bug if we use KeyType and KeyTypePointer here.

https://bugzilla.mozilla.org/show_bug.cgi?id=1597715

And I also had to apply this bug because the SunOS libc didn't like the undefined behavior they were using here. I'm not sure if this would be an issue on other operating systems.

This probably needs to be split out into three different commits, I'm just writing this note to you so you can see what I did and use the information for yourself to decide how best to proceed.

Good luck, G4JC.