Issue #1629 - Alternative idea for how to address <link> disabled.
I got really frustrated while trying to figure out Emilio's code and happened to come up with this alternative approach while working on it. It doesn't match the spec perfectly, but it produces the same visual behavior as the spec in my testing.

Basically, this differs from the official implementation by simply asserting that any <link> element that has a disabled attribute is actually an alternate stylesheet, even if it normally wouldn't meet the standard criteria for being one. Because alternate stylesheets are already disabled by default, declaring a sheet to be alternate and giving it a title means it isn't loaded by default.

This is a very primitive implementation and likely isn't suitable for every website out there, probably a better version of this code would have to (at bare minimum) be able to generate a unique name for every untitled disabled stylesheet, like Untitled 1, Untitled 2, etc. And if any site ever actually depends on disabled stylesheets being removed from document.styleSheets, this implementation would fail.